Refactored gallery
